+++ instrumentation-loongsuite/loongsuite-instrumentation-dashscope/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fix extraction of `gen_ai.output.messages` for message-format text responses
+  that omit optional `tool_calls`.
+
 ## Version 0.5.0 (2026-05-11)
 
 There are no changelog entries for this release.